+++ sdk/search/search-documents/test/public/node/searchClient.spec.ts
@@ -19,6 +19,7 @@ import {
 } from "../../../src";
 import { SearchFieldArray, SelectArray } from "../../../src/indexModels";
 import { delay, serviceVersions } from "../../../src/serviceUtils";
+import { COMPRESSION_DISABLED } from "../../compressionDisabled";
 import { Hotel } from "../utils/interfaces";
 import { createClients } from "../utils/recordedClient";
 import { createIndex, createRandomIndexName, populateIndex, WAIT_TIME } from "../utils/setup";
@@ -379,7 +380,7 @@ versionsToTest(serviceVersions, {}, (serviceVersion, onVersions) => {
     });
   });
 
-  onVersions({ minVer: "2023-10-01-Preview" }).describe(
+  onVersions({ minVer: "2024-03-01-Preview" }).describe(
     "SearchClient tests",
     function (this: Suite) {
       let recorder: Recorder;
@@ -459,13 +460,13 @@ versionsToTest(serviceVersions, {}, (serviceVersion, onVersions) => {
                   keywordFields: [
                     {
                       name: "tags",
-                      state: "unused",
+                      state: "used",
                     },
                   ],
                   rerankerInput: {
                     content:
                       "Best hotel in town if you like luxury hotels. They have an amazing infinity pool, a spa, and a really helpful concierge. The location is perfect -- right downtown, close to all the tourist attractions. We highly recommend this hotel.",
-                    keywords: "",
+                    keywords: "pool\r\nview\r\nwifi\r\nconcierge",
                     title: "Fancy Stay",
                   },
                   titleField: {
@@ -496,7 +497,7 @@ versionsToTest(serviceVersions, {}, (serviceVersion, onVersions) => {
         for await (const result of searchResults.results) {
           resultIds.push(result.document.hotelId);
         }
-        assert.deepEqual(["3", "9", "1"], resultIds);
+        assert.deepEqual(["1", "9", "3"], resultIds);
       });
 
       it("search with semantic error handling", async function () {
@@ -590,6 +591,44 @@ versionsToTest(serviceVersions, {}, (serviceVersion, onVersions) => {
         }
         assert.deepEqual(["1", "3", "4"], resultIds);
       });
+
+      it("oversampling compressed vectors", async function () {
+        // This live test is disabled due to temporary limitations with the new OpenAI service
+        if (isLiveMode()) {
+          this.skip();
+        }
+        // Currently unable to create a compression resource
+        if (COMPRESSION_DISABLED) {
+          this.skip();
+        }
+        const embeddings = await openAIClient.getEmbeddings(
+          env.AZURE_OPENAI_DEPLOYMENT_NAME ?? "deployment-name",
+          ["What are the most luxurious hotels?"],
+        );
+
+        const embedding = embeddings.data[0].embedding;
+        const searchResults = await searchClient.search("*", {
+          vectorSearchOptions: {
+            queries: [
+              {
+                kind: "vector",
+                vector: embedding,
+                kNearestNeighborsCount: 3,
+                fields: ["compressedVectorDescription"],
+                oversampling: 2,
+              },
+            ],
+          },
+          top: 3,
+          select: ["hotelId"],
+        });
+
+        const resultIds = [];
+        for await (const result of searchResults.results) {
+          resultIds.push(result.document.hotelId);
+        }
+        assert.deepEqual(["1", "3", "4"], resultIds);
+      });
     },
   );
 });
@@ -617,8 +656,8 @@ versionsToTest(serviceVersions, {}, (serviceVersion, onVersions) => {
 
       it("defaults to the current apiVersion", () => {
         const client = new SearchClient<Hotel>("", "", credential);
-        assert.equal("2023-10-01-Preview", client.serviceVersion);
-        assert.equal("2023-10-01-Preview", client.apiVersion);
+        assert.equal("2024-03-01-Preview", client.serviceVersion);
+        assert.equal("2024-03-01-Preview", client.apiVersion);
       });
     });
   });